ajout des autorisations

+++ app/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import get_object_or_404
 from django.shortcuts import redirect, render
 from django.contrib.auth.decorators import login_required
+from django.conf import settings
+from django.http import Http404
 
 from . import forms, models
 from authentication.models import User
@@ -16,14 +18,19 @@ def home(request):
 
 """Vue de la création de ticket"""
 @login_required
+
+# @permission_required('app.add_ticket', raise_exception=True)
 def ticket_creator_form(request):
     form = forms.TicketForm()
     if request.method == 'POST':
         form = forms.TicketForm(request.POST, request.FILES)
         if form.is_valid():
             ticket = form.save(commit=False)
+
             ticket.uploader = request.user
             ticket.save()
+            request.user.ticket_uploaded = ticket
+            print("___________________________________",request.user.ticket_uploaded == ticket, "___________________________________")
             return redirect('home')
     return render(request, 'app/ticket_creator.html', context={'form': form})
 
@@ -62,21 +69,24 @@ def edit_ticket(request, ticket_id):
     ticket = get_object_or_404(models.Ticket, id=ticket_id)
     edit_form = forms.TicketForm(instance=ticket)
     delete_form = forms.DeleteTicketForm()
-    if request.method == 'POST':
-        if 'edit_ticket' in request.POST:
-            edit_form = forms.TicketForm(request.POST, instance=ticket)
-            if edit_form.is_valid():
-                edit_form.save()
-                return redirect('home')
-        if 'delete_ticket' in request.POST:
-            delete_form = forms.DeleteTicketForm(request.POST)
-            if delete_form.is_valid():
-                ticket.delete()
-                return redirect('home')
-    context = {
-                'edit_form': edit_form,
-                'delete_form': delete_form
-              }
+    if ticket.uploader == request.user:
+        if request.method == 'POST':
+            if 'edit_ticket' in request.POST:
+                edit_form = forms.TicketForm(request.POST, instance=ticket)
+                if edit_form.is_valid():
+                    edit_form.save()
+                    return redirect('home')
+            if 'delete_ticket' in request.POST:
+                delete_form = forms.DeleteTicketForm(request.POST)
+                if delete_form.is_valid():
+                    ticket.delete()
+                    return redirect('home')
+        context = {
+                    'edit_form': edit_form,
+                    'delete_form': delete_form
+                }
+    else:
+        raise Http404
     return render(request, 'app/edit_ticket.html', context=context)
 
 
@@ -86,21 +96,24 @@ def edit_review(request, review_id):
     review = get_object_or_404(models.Review, id=review_id)
     edit_form = forms.ReviewForm(instance=review)
     delete_form = forms.DeleteReviewForm()
-    if request.method == 'POST':
-        if 'edit_review' in request.POST:
-            edit_form = forms.ReviewForm(request.POST, instance=review)
-            if edit_form.is_valid():
-                edit_form.save()
-                return redirect('home')
-        if 'delete_review' in request.POST:
-            delete_form = forms.DeleteReviewForm(request.POST)
-            if delete_form.is_valid():
-                review.delete()
-                return redirect('home')
-    context = {
-                'edit_form': edit_form,
-                'delete_form': delete_form
-              }
+    if review.uploader == request.user:
+        if request.method == 'POST':
+            if 'edit_review' in request.POST:
+                edit_form = forms.ReviewForm(request.POST, instance=review)
+                if edit_form.is_valid():
+                    edit_form.save()
+                    return redirect('home')
+            if 'delete_review' in request.POST:
+                delete_form = forms.DeleteReviewForm(request.POST)
+                if delete_form.is_valid():
+                    review.delete()
+                    return redirect('home')
+        context = {
+                    'edit_form': edit_form,
+                    'delete_form': delete_form
+                }
+    else:
+        raise Http404
     return render(request, 'app/edit_review.html', context=context)
